backend/requirements.txt
alembic==1.7.5
aniso8601==6.0.0
astroid==2.2.5
atomicwrites==1.4.0
attrs==21.2.0
awscli==1.22.13
Babel==2.9.0
botocore==1.23.13
Click==7.0
colorama==0.4.3
docutils==0.15.2
ecdsa==0.13.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.17.1
greenlet==1.1.2
gunicorn==20.0.4
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.0
Mako==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==0.13.1
psycopg2-binary==2.8.2
py==1.11.0
pyasn1==0.4.8
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==3.0.6
pytest==6.2.2
python-dateutil==2.6.0
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.12.0
SQLAlchemy==1.4.3
toml==0.10.2
typed-ast==1.4.2
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==0.15.6
wrapt==1.11.1
WTForms==3.0.0
zipp==3.6.0